Update to doctest 2.4.0
While making this update, it made sense to consolidate external
dependencies in a central location. That's why I decided to move
the call to FetchContent to the root CMake file.

It also made sense to only perform a shallow copy of the repo.

Relevant Links:
- Issue 10: Update to doctest 2.4.0
(https://github.com/tybl/libvodka/issues/10)