Compile tests when root project
I thought I had a cool system to prevent accidentally enabling tests
as a subproject. It used a variable that I thought was set depending
on whether the current project name matched the root project name.

Well, the variable wasn't getting set correctly. Instead of evaluating
the current and root project names, it was just assigning the variable
names, as a semicolon separated list, to the variable. CMake said,
"if variable? Yup, looks like variable is set. Let do this thing!"
Well, it said that a lot. Because it was always set.

So, I fixed that by figuring out the correct way of conditionally
assigning a variable in CMake. Cool.

But it wasn't. Nothing changed. The variable was still assigned a
wonky value. I said I was importing this library into one of my own
libraries, right? Yeah, well guess what variable I set in that
project? That's right! The same one! Using the same broken logic!
I'll just go fix-- Wait a minute! My root project is setting the
variable my imported projects use to determine if they are root
projects?? That's not gunna work!

I have decided that user defined CMake variables are the devil and
shouldn't be used. Instead, I evaluate the current and root project
names at each location that is dependent.

That was a fun head scratcher!

I also had to guard the compiler_settings library and the FetchContent
for doctest behind the same check and figured I shouldn't force a
project to use linters if they didn't ask for them. After all that,
guess what the importing project gets? Just the library they asked
for! I will need to fill out the target features for the library
though.

Relevant Links:
- How to do conditional variable assignment with cmake?
(https://stackoverflow.com/questions/51320553)
- Issue 17: Compile tests when root project
(https://github.com/tybl/libvodka/issues/17)