Remove directory names from namespaces
The namespaces are a bit onerous.

I think using my github user name as a root namespace is a good
practice. The supporting side of that decision is discussed by
foonathan below. His practice also includes a nested namespace for
the library name, which makes sense too. However, yesterday I thought
it would be a good idea to rename my library projects to begin with
'lib', but I don't want to have an additional 'lib' in the namespaces.
I think I can safely ignore the 'lib' part of the project name as long
as I don't create duplicate projects with and without 'lib' in the
name.

Then to add to the madness, I like to include the directory structure
in the namespace. And since I am trying to separate the standard
library components into their individual files, under the directory
name that originally included them, the directory structure could
become expansive.

Using the directory structure in the namespace is probably overkill
though and can safely be ignored. Besides, this library implements a
subset of the standard library and most of that is in a flat
namespace. Let's start with removing the directory structure from the
namespace and see how accessing the library feels from there.

Relevant Links:
- One of C++ most underrated features: Namespace aliases
(https://foonathan.net/2016/01/namespace-alias/)
- Issue 20: Lessen namespace depth
(https://github.com/tybl/libvodka/issues/20)